Moved much code from assembler to other places, many crutches. But program works for now
Moved much code from assembler to other places, many crutches. But program works for now
