+++ cocas/targets/cdm8e/__init__.py
@@ -1,0 +1,2 @@
+from .code_segments import CodeSegments
+from .target_instructions import TargetInstructions
